Updated landing page, added login page, stylesheet
The landing page has been updated with new images that are license-free to avoid copyright infringement.

A new page has been added for the user to sign in to their account. On this page, the user can also fill up a registration form if they do not have an account yet.

In order to maintain style consistency, an external stylesheet has been added that will be used for all pages.